--- app/src/main/java/com/example/moonlight/ui/categoria/CategoriaViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.moonlight.ui.categoria
-
-import androidx.lifecycle.ViewModel
-
-class CategoriaViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
