+++ app/src/main/java/com/example/moonlight/ui/category/CategoryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moonlight.ui.categoria
+package com.example.moonlight.ui.category
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -8,24 +8,24 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.moonlight.R
 
-class CategoriaFragment : Fragment() {
+class CategoryFragment : Fragment() {
 
     companion object {
-        fun newInstance() = CategoriaFragment()
+        fun newInstance() = CategoryFragment()
     }
 
-    private lateinit var viewModel: CategoriaViewModel
+    private lateinit var viewModel: CategoryViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_categoria, container, false)
+        return inflater.inflate(R.layout.fragment_category, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(CategoriaViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(CategoryViewModel::class.java)
         // TODO: Use the ViewModel
     }
