+++ app/src/main/java/com/example/moonlight/ui/category/CategoryViewModel.kt
@@ -1,0 +1,7 @@
+package com.example.moonlight.ui.category
+
+import androidx.lifecycle.ViewModel
+
+class CategoryViewModel : ViewModel() {
+    // TODO: Implement the ViewModel
+}
