+++ app/src/main/res/layout/fragment_category.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.categoria.CategoriaFragment">
+    tools:context=".ui.category.CategoryFragment">
 
     <TextView
         android:layout_width="match_parent"
